+++ testing/renovation/platforms/declaration/helpers/getComponentOptions.ts
